requirements.txt
apache-flink==1.17.1
apache-kafka-python==2.0.2
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
prometheus-client==0.17.1
mlflow==2.7.1
pyflink-ml==1.17.1 